port benchmarks.
on the 4M case, this setup seems to do roughly 4% faster both ways (meh) and score about 3x better on bytes/op on decryption and 2x on encryption (not surprising, since one round of buffering went away since we can stream outputs with this format instead of sitting on them until we have the MAC to pass into ASN1 and *then* start outputting body).

more like 20% faster for the 1k case.  presumably mostly again because the overhead of ASN1 and that extra buffer disappeared.

overall, performance wasn't so much the reason for these format changes, but not bad improvements nonetheless.  the cipher still can't operate fully streamwise on very large (i.e., multi-gigabyte) files, but the format also no longer poses major hurdles to that in the future; we just have to fix the encryptor to be willing to do two passes on the input.